States Machine feat & NavMesh feat
Used @JossPabloPC States Machine scripts for the enemy behaviour. It has an Idle state and a ChasePlayer state. It enters Idle bu default and when the player enters its trigger of field of view it will change to the chase player state, when the player exits this collider it will return to idle.

Also implemented the navmesh so that it moves towards the "player" in a given area.