AttackPlayer State feat
Created a new state called AttackPlayer that will manage the attack of the enemy to the player.
It attacks every "attackSpan" seconds, this is managed with a coroutine.
The enemy will enter this state if it is less than 1.4 of distance with the player, and will chase the player again if the distance between them is greater than 1.4.